cli app v.0 with postprocessing in main and head-only model
