+++ .gitignore
@@ -1,5 +1,5 @@
 .DS_Store
 /data/raw
 /notebooks/model
-/model_files
+/models
 __pycache__/
